Implement smart wrapping
When the window is smaller than the body wrap setting, the textview is set
to wrap word and Commit will wrap to the requested width on save.

When the window is bigger than the body wrap setting, the textview wrap is
disabled and Commit will not wrap the text on save.

This should provide a great experience on mobile and other small screens as
well as allowing developers to disable wrapping when needed.

The right margin / ruler always shows, so even if the warpping is disabled,
there is still an indication as to where the text is preferred to break.

                                                       _.- ~~^^^'~- _ __ .,.- ~ ~ ~  ~  -. _
                                             ________,'       ::.                       _,-  ~ -.
                                            ((      ~_\   -s-  ::                     ,'          ;,
                                             \\       <.._ .;;;`                     ;           }  `',
                                              ``======='    _ _- _ (   }             `,          ,'\,  `,
                                                           ((/ _ _,i   ! _ ~ - -- - _ _'_-_,_,,,'    \,  ;
                                                   cfbd       ((((____/            (,(,(, ____>        \,'